Move .example.tsx files
