+++ packages/website/docs/authored/diagram/node-description/node-description-docs.tsx
@@ -1,6 +1,6 @@
 import { NodeDescription, NodePanel } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./node-description.example.jsx';
+import exampleCode from '!!raw-loader!@site/docs/code-examples/node-description.example.tsx';
 
 export function Docs() {
   return (
