--- packages/website/docs/authored/diagram/node-icon/node-icon.example.jsx
@@ -1,10 +1,0 @@
-function Page() {
-  return (
-    <NodePanel.Root>
-      <NodePanel.Header>
-        <NodeIcon icon={<User />} />
-        Node with Icon
-      </NodePanel.Header>
-    </NodePanel.Root>
-  );
-}
