+++ packages/website/docs/authored/ui/input/input-preview/input-with-start-adornment.tsx
