+++ packages/website/docs/authored/ui/input/input-preview/simple-input.tsx
