+++ packages/website/docs/authored/ui/menu/menu-docs.tsx
@@ -1,6 +1,6 @@
 import { Menu, LabelButton } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./menu.example.jsx';
+import exampleCode from '!!raw-loader!@site/docs/code-examples/menu.example.tsx';
 
 export function MenuDocs() {
   return (
