+++ packages/website/docs/authored/ui/radio/radio-docs.tsx
@@ -1,6 +1,6 @@
 import { Radio } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./radio.example.jsx';
+import exampleCode from '!!raw-loader!@site/docs/code-examples/radio.example.tsx';
 import React from 'react';
 
 export function RadioDocs() {
