+++ packages/website/docs/authored/ui/select/select-docs.tsx
@@ -1,6 +1,6 @@
 import { Select } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./select.example.jsx';
+import exampleCode from '!!raw-loader!@site/docs/code-examples/select.example.tsx';
 
 export function SelectDocs() {
   return (
