--- packages/website/docs/authored/ui/switch/switch.example.jsx
@@ -1,9 +1,0 @@
-function Page() {
-  return (
-    <Switch
-      size="medium"
-      disabled={false}
-      onChange={() => {}}
-    />
-  );
-}
