+++ packages/website/docs/authored/ui/tooltip/tooltip-docs.tsx
@@ -1,6 +1,6 @@
 import { Tooltip } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./tooltip.example.jsx';
+import exampleCode from '!!raw-loader!@site/docs/code-examples/tooltip.example.tsx';
 
 export function TooltipDocs() {
   return (
