+++ packages/website/docs/authored/ui/tooltip/tooltip-trigger/tooltip-trigger-docs.tsx
@@ -1,6 +1,6 @@
 import { Tooltip } from '@synergycodes/axiom';
 import { ComponentPage } from '@site/src/components/component-utils/component-page/component-page';
-import exampleCode from '!!raw-loader!./tooltip-trigger.example.jsx';
+import exampleCode from '!!raw-loader!@site/docs/code-examples/tooltip-trigger.example.tsx';
 
 export function TooltipTriggerDocs() {
   return (
