+++ packages/website/docs/code-examples/avatar.example.tsx
@@ -1,4 +1,4 @@
-function Page() {
+function AvatarExample() {
   return (
     <Avatar
       imageUrl="https://picsum.photos/id/237/200/300"
