+++ packages/website/docs/code-examples/input.example.tsx
@@ -1,3 +1,3 @@
-function Page() {
+function InputExample() {
   return <Input defaultValue="Hello World" size="medium" />;
 }
