+++ packages/website/docs/code-examples/menu.example.tsx
@@ -1,4 +1,4 @@
-function Page() {
+function MenuExample() {
   return (
     <Menu items={[{ label: 'Star Wars' }, { label: 'Star Treck' }]}>
       <LabelButton label="Open" />
