+++ packages/website/docs/code-examples/node-description.example.tsx
@@ -1,0 +1,9 @@
+function NodeDescriptionExample() {
+  return (
+    <NodePanel.Root selected={false}>
+      <NodePanel.Header>
+        <NodeDescription label="Node" description="and some description" />
+      </NodePanel.Header>
+    </NodePanel.Root>
+  );
+}
