+++ packages/website/docs/code-examples/switch.example.tsx
@@ -1,0 +1,3 @@
+function SwitchExample() {
+  return <Switch size="medium" disabled={false} onChange={() => {}} />;
+}
