+++ packages/website/docs/code-examples/tooltip-content.example.tsx
@@ -1,11 +1,11 @@
-function Page() {
+function TooltipContentExample() {
   return (
-    <Tooltip>
+    <Tooltip open={true}>
       <Tooltip.Trigger>
         <span>Tooltip</span>
       </Tooltip.Trigger>
       <Tooltip.Content>
-        <span>Tooltip</span>
+        <span>Tooltip Content Component</span>
       </Tooltip.Content>
     </Tooltip>
   );
