+++ packages/website/docs/code-examples/tooltip-trigger.example.tsx
@@ -1,4 +1,4 @@
-function Page() {
+function TooltipTriggerExample() {
   return (
     <Tooltip>
       <Tooltip.Trigger>
